Implemented PageController.java
